tests/requirements.txt
PyYaml
GitPython
